fix: BadgeController의 getAllBadgeList함수 뱃지가 없는 경우 NO_CONTENT를 반환하는 것 수정,...

See merge request s08-webmobile1-sub2/S08P12A107!131

+++ a107/src/main/java/com/ssafy/a107/api/service/BadgeServiceImpl.java
@@ -30,7 +30,8 @@ public class BadgeServiceImpl implements BadgeService {
 
 
     @Override
-    public List<BadgeRes> getBadgeByUserSeq(Long userSeq) {
+    public List<BadgeRes> getBadgeByUserSeq(Long userSeq) throws NotFoundException{
+        if(userRepository.existsById(userSeq)) throw new NotFoundException("Wrong User Seq!");
         return badgeRepository.findBadgesByUserSeq(userSeq).stream()
                 .map(BadgeRes::new)
                 .collect(Collectors.toList());
@@ -79,8 +80,10 @@ public class BadgeServiceImpl implements BadgeService {
     }
 
     @Override
-    public void deleteBadge(Long seq) {
-        badgeRepository.deleteById(seq);
+    public void deleteBadge(Long seq) throws NotFoundException{
+        if(badgeRepository.existsById(seq)){
+            badgeRepository.deleteById(seq);
+        }else throw new NotFoundException("Invalid badge sequence!");
     }
